# path: performance and stability improvements on all platforms