--- a/benchmark/path/extname-posix.js
+++ b/benchmark/path/extname-posix.js
@@ -3,24 +3,35 @@ var path = require('path');
 var v8 = require('v8');
 
 var bench = common.createBenchmark(main, {
-  type: ['win32', 'posix'],
-  n: [1e6],
+  path: [
+    '',
+    '/',
+    '/foo',
+    'foo/.bar.baz',
+    'index.html',
+    'index',
+    'foo/bar/..baz.quux',
+    'foo/bar/...baz.quux',
+    '/foo/bar/baz/asdf/quux',
+    '/foo/bar/baz/asdf/quux.foobarbazasdfquux'
+  ],
+  n: [1e6]
 });
 
 function main(conf) {
   var n = +conf.n;
-  var p = path[conf.type];
+  var p = path.posix;
+  var input = '' + conf.path;
 
   // Force optimization before starting the benchmark
-  p.extname('index.html');
+  p.extname(input);
   v8.setFlagsFromString('--allow_natives_syntax');
   eval('%OptimizeFunctionOnNextCall(p.extname)');
-  p.extname('index.html');
+  p.extname(input);
 
   bench.start();
   for (var i = 0; i < n; i++) {
-    p.extname('index.html');
-    p.extname('index');
+    p.extname(input);
   }
   bench.end(n);
 }