--- a/benchmark/path/format-win32.js
+++ b/benchmark/path/format-win32.js
@@ -3,36 +3,33 @@ var path = require('path');
 var v8 = require('v8');
 
 var bench = common.createBenchmark(main, {
-  type: ['win32', 'posix'],
-  n: [1e7],
+  props: [
+    ['C:\\', 'C:\\path\\dir', 'index.html', '.html', 'index'].join('|')
+  ],
+  n: [1e7]
 });
 
 function main(conf) {
   var n = +conf.n;
-  var p = path[conf.type];
-  var test = conf.type === 'win32' ? {
-    root: 'C:\\',
-    dir: 'C:\\path\\dir',
-    base: 'index.html',
-    ext: '.html',
-    name: 'index'
-  } : {
-    root : '/',
-    dir : '/home/user/dir',
-    base : 'index.html',
-    ext : '.html',
-    name : 'index'
+  var p = path.win32;
+  var props = ('' + conf.props).split('|');
+  var obj = {
+    root: props[0] || '',
+    dir: props[1] || '',
+    base: props[2] || '',
+    ext: props[3] || '',
+    name: props[4] || '',
   };
 
   // Force optimization before starting the benchmark
-  p.format(test);
+  p.format(obj);
   v8.setFlagsFromString('--allow_natives_syntax');
   eval('%OptimizeFunctionOnNextCall(p.format)');
-  p.format(test);
+  p.format(obj);
 
   bench.start();
   for (var i = 0; i < n; i++) {
-    p.format(test);
+    p.format(obj);
   }
   bench.end(n);
 }