--- a/benchmark/path/parse-posix.js
+++ b/benchmark/path/parse-posix.js
@@ -1,0 +1,34 @@
+var common = require('../common.js');
+var path = require('path');
+var v8 = require('v8');
+
+var bench = common.createBenchmark(main, {
+  path: [
+    '',
+    '/',
+    '/foo',
+    '/foo/bar.baz',
+    'foo/.bar.baz',
+    'foo/bar',
+    '/foo/bar/baz/asdf/.quux'
+  ],
+  n: [1e6]
+});
+
+function main(conf) {
+  var n = +conf.n;
+  var p = path.posix;
+  var input = '' + conf.path;
+
+  // Force optimization before starting the benchmark
+  p.parse(input);
+  v8.setFlagsFromString('--allow_natives_syntax');
+  eval('%OptimizeFunctionOnNextCall(p.parse)');
+  p.parse(input);
+
+  bench.start();
+  for (var i = 0; i < n; i++) {
+    p.parse(input);
+  }
+  bench.end(n);
+}