--- a/benchmark/path/parse-win32.js
+++ b/benchmark/path/parse-win32.js
@@ -3,26 +3,33 @@ var path = require('path');
 var v8 = require('v8');
 
 var bench = common.createBenchmark(main, {
-  type: ['win32', 'posix'],
-  n: [1e6],
+  path: [
+    '',
+    'C:\\',
+    'C:\\foo',
+    '\\foo',
+    'E:\\foo\\bar.baz',
+    'foo\\.bar.baz',
+    'foo\\bar',
+    '\\foo\\bar\\baz\\asdf\\.quux'
+  ],
+  n: [1e6]
 });
 
 function main(conf) {
   var n = +conf.n;
-  var p = path[conf.type];
-  var test = conf.type === 'win32'
-    ? 'C:\\path\\dir\\index.html'
-    : '/home/user/dir/index.html';
+  var p = path.win32;
+  var input = '' + conf.path;
 
   // Force optimization before starting the benchmark
-  p.parse(test);
+  p.parse(input);
   v8.setFlagsFromString('--allow_natives_syntax');
   eval('%OptimizeFunctionOnNextCall(p.parse)');
-  p.parse(test);
+  p.parse(input);
 
   bench.start();
   for (var i = 0; i < n; i++) {
-    p.parse(test);
+    p.parse(input);
   }
   bench.end(n);
 }