--- a/test/parallel/test-path-zero-length-strings.js
+++ b/test/parallel/test-path-zero-length-strings.js
@@ -12,17 +12,21 @@ const pwd = process.cwd();
 
 // join will internally ignore all the zero-length strings and it will return
 // '.' if the joined string is a zero-length string.
-assert.equal(path.join(''), '.');
-assert.equal(path.join('', ''), '.');
+assert.equal(path.posix.join(''), '.');
+assert.equal(path.posix.join('', ''), '.');
+assert.equal(path.win32.join(''), '.');
+assert.equal(path.win32.join('', ''), '.');
 assert.equal(path.join(pwd), pwd);
 assert.equal(path.join(pwd, ''), pwd);
 
 // normalize will return '.' if the input is a zero-length string
-assert.equal(path.normalize(''), '.');
+assert.equal(path.posix.normalize(''), '.');
+assert.equal(path.win32.normalize(''), '.');
 assert.equal(path.normalize(pwd), pwd);
 
 // Since '' is not a valid path in any of the common environments, return false
-assert.equal(path.isAbsolute(''), false);
+assert.equal(path.posix.isAbsolute(''), false);
+assert.equal(path.win32.isAbsolute(''), false);
 
 // resolve, internally ignores all the zero-length strings and returns the
 // current working directory