@evanlucas 2016-02-06
Wow. Thanks @mscdex. I wonder if this brings any performance improvements to require as well.