@mscdex 2016-02-06
@evanlucas I just ran the `misc/module-loader` benchmark several times and I see ~17-35% improvement there.