@jasnell 2016-02-07
Marking this don't land on v4 for now... this is definitely one that we'd want to prove out for a while before backporting.