@mathiasbynens 2016-02-08
:+1: 

> these results were obtained from a benchmark.js-based benchmark runner I'm working on

I’d be interested in the runner as well. Is it open-source?