@rvagg 2016-02-15
@mscdex I've just backported this to v5.x which was a little bit tricky mainly because of the error message changes (arg names in `"` continues to be a hassle). If you wouldn't mind having a look at 4dae8ca to see if it's been done satisfactorily I'd appreciate the help.

One thing that stood out to me when doing this was the removal of the error messages around `pathObject.root`, can you explain why these were removed and why they were removed from the tests and how this doesn't constitute a semver-major change? It seems to me that if we're no longer able to cause an error of some kind then it may warrant holding this off until the next major.