@mscdex 2016-02-15
@rvagg As far as the actual backport goes, it LGTM, the only differences are in the error strings as you noted.

 The particular error removal and test changes were not made by me, those were introduced back in d1000b413793 which apparently never made it into v5.x. Also FWIW d1000b413793 adds some additional `path` test cases that v5.x is lacking.