@TheAlphaNerd 2016-02-15
Oh no's... it would appear that this change is breaking the test suite for ESlint and for stylus

https://github.com/nodejs/citgm/issues/81
https://github.com/nodejs/citgm/issues/80

Specifically it seems to be instances where path should be returning a string, but is now returning undefined.

If this is intentional than this PR should be semver-major. If it is not intentional than we need to fix it. I'll take a peak and see if I can get a test up and running