@mscdex 2016-02-15
It's not intentional, we were just lacking tests for certain cases (previously `regexp.exec()` was implicitly converting its inputs to string). I will submit a PR to add them soon.

However, at some point we should probably disallow non-string inputs (`semver-major`) for the problematic path functions because IMHO passing `null`, `undefined`, `{}`, should not result in path component information that contains `'null'`, `'undefined'`, or `'[object Object]'` respectively. That seems like unexpected behavior.