@TheAlphaNerd 2016-02-15
@mscdex thanks for clarifying

@rvagg in light of this I think we should back all the commits from this PR out of v5.x