@mscdex 2016-02-15
I think the `semver-major` label on this can be removed with the caveat that #5244 would need to land with it?