@rvagg 2016-02-15
@TheAlphaNerd @mscdex regarding v5.x, I think my preference would still be to move ahead here as this is not supposed to be a breaking change and it would be a shame to lump all of the "this may find edge cases" changes into v6 and then be scrambling for a few weeks to clean them all up.
If we're not confident enough to put this in v5 then we probably shouldn't even be putting it in v6. Is there a way to increase that confidence or are we more comfortable now with the citgm run and #5244 fix?