@TheAlphaNerd 2016-02-15
@rvagg I have no problem with that myself. The patch was not landing cleanly on v5.x though, so perhaps we want to open a second PR for that?