@rvagg 2016-02-15
Cherry pick is fine, it's just these annoying error message format updates that have messed things up, manually working around them is not hard though so no need for another PR, I'll deal with it when it lands