@TheAlphaNerd 2016-02-15
ah ok cool. I'm multi-tasking like mad now and was just doing sanity checks without diving to far in. Thanks for taking the time