@rvagg 2016-02-18
removed semver-major label