@eggggger 2016-02-25
v5.7
`> path.relative('app/index.js', '.')
'../../../yadan'`
v5.6
`> path.relative('app/index.js', '.')
'../..'`
The result is different,why?