@evanlucas 2016-02-25
@eggggger that is fixed by https://github.com/nodejs/node/pull/5389. A release with the fix should be going out soon. Thanks!